test(orders): all orders test have been adjusted
All tests now passed, and all relevant test have been written

this resolves #5
